chore(ui): refactor
