+++ ui/src/override/components/namespaces/useTabs.ts
@@ -1,6 +1,7 @@
-import {Component} from "vue";
 import {useRoute} from "vue-router";
 import {useI18n} from "vue-i18n";
+
+import {Tab, ORDER} from "../../../components/namespaces/utils/useHelpers";
 
 import DemoNamespace from "../../../components/demo/Namespace.vue";
 
@@ -12,37 +13,6 @@ import Executions from "../../../components/executions/Executions.vue";
 import Dependencies from "../../../components/namespaces/components/content/Dependencies.vue";
 import KVTable from "../../../components/kv/KVTable.vue";
 import EditorView from "../../../components/inputs/EditorView.vue";
-
-interface Tab {
-    locked?: boolean;
-
-    name: string;
-    title: string;
-    component: Component;
-
-    props?: Record<string, any>;
-}
-
-interface Details {
-    title: string;
-    breadcrumb: Record<string, any>[];
-}
-
-const ORDER = [
-    "blueprints",
-    "overview",
-    "edit",
-    "flows",
-    "executions",
-    "dependencies",
-    "secrets",
-    "variables",
-    "plugin-defaults",
-    "kv",
-    "files",
-    "history",
-    "audit-logs",
-];
 
 const lockedProps = (tab: string) => ({
     locked: true,
@@ -141,21 +111,5 @@ export function useTabs() {
     // Ensure the order of tabs is following the ORDER array
     tabs.sort((a, b) => ORDER.indexOf(a.name) - ORDER.indexOf(b.name));
 
-    const parts = namespace.split(".") ?? [];
-    const details: Details = {
-        title: parts.at(-1) || t("namespaces"),
-        breadcrumb: [
-            {label: t("namespaces"), link: {name: "namespaces/list"}},
-            ...parts.map((_: string, index: number) => ({
-                label: parts[index],
-                link: {
-                    name: "namespaces/update",
-                    params: {id: parts.slice(0, index + 1).join(".")},
-                },
-                disabled: index === parts.length - 1,
-            })),
-        ],
-    };
-
-    return {tabs, details};
+    return {tabs};
 }
